+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file. It follows 
 ### Added
 - "news" argument added for returning latest Pitchfork news (up to the 10 latest articles)
 - Alternate flags for most flags (see `pitchfork help` for details)
+- Separated elements for extracting data from Pitchfork.com into "pitchfork" package
 
 ## 0.1.0 - 2015-01-25
 ### Added